CASMUSER-2810 Implement job based UAIs with timeouts and one_shot
