+++ swagger_server/uas_lib/uas_base.py
@@ -25,14 +25,9 @@ Base Class for User Access Service Operations
 Copyright 2020 Hewlett Packard Enterprise Development LP
 """
 
-import os
-import json
-import uuid
 import time
 from datetime import datetime, timezone
 from flask import abort
-import sshpubkeys
-import sshpubkeys.exceptions as sshExceptions
 from kubernetes import config, client
 from kubernetes.client.rest import ApiException
 from kubernetes.client import Configuration
@@ -40,18 +35,9 @@ from kubernetes.client.api import core_v1_api
 from swagger_server.uas_lib.uas_logging import logger
 from swagger_server.models import UAI
 from swagger_server.uas_lib.uas_cfg import UasCfg
-from swagger_server.uas_data_model.uai_resource import UAIResource
-from swagger_server.uas_data_model.uai_image import UAIImage
-# For now because the testing seems to need it, pull in UAS_AUTH_LOGGER
-from swagger_server.uas_lib.uas_auth import UAS_AUTH_LOGGER
 
 # picking 40 seconds so that it's under the gateway timeout
 UAI_IP_TIMEOUT = 40
-
-# All UAIs have the following toleration to allow them to run
-# on nodes that are tainted against non-UAI activity.  The list
-# can be extended using UAI Class toleration lists.
-BASE_UAI_TOLERATIONS = [client.V1Toleration(key="uai_only", operator="Exists")]
 
 
 class UasBase:
@@ -70,7 +56,7 @@ class UasBase:
             k8s_config.assert_hostname = False
         Configuration.set_default(k8s_config)
         self.api = core_v1_api.CoreV1Api()
-        self.apps_v1 = client.AppsV1Api()
+        self.batch_v1 = client.BatchV1Api()
         self.uas_cfg = UasCfg()
 
     @staticmethod
@@ -184,7 +170,7 @@ class UasBase:
             )
         except ApiException as err:
             # if we get 404 we don't want to abort because it's possible that
-            # other parts are still laying around (deployment for example)
+            # other parts are still laying around (job for example)
             if err.status != 404:
                 logger.error(
                     "Failed to delete service %s: %s",
@@ -199,48 +185,49 @@ class UasBase:
                 )
         return resp
 
-    def create_deployment(self, deployment, namespace):
-        """Create a UAI deployment
+    def create_job(self, job, namespace):
+        """Create a UAI job
 
         """
         resp = None
         try:
             logger.info(
-                "creating deployment %s in namespace %s",
-                deployment.metadata.name,
+                "creating job %s in namespace %s",
+                job.metadata.name,
                 namespace
             )
-            resp = self.apps_v1.create_namespaced_deployment(
-                body=deployment,
+            resp = self.batch_v1.create_namespaced_job(
+                body=job,
                 namespace=namespace
             )
         except ApiException as err:
             logger.error(
-                "Failed to create deployment %s: %s",
-                deployment.metadata.name,
+                "Failed to create job %s: %s",
+                job.metadata.name,
                 err.reason
             )
+            logger.debug("namespace = %s, job = \n%s", namespace, job)
             abort(
                 err.status,
-                "Failed to create deployment %s: %s" % (
-                    deployment.metadata.name, err.reason
+                "Failed to create job %s: %s" % (
+                    job.metadata.name, err.reason
                 )
             )
         return resp
 
-    def delete_deployment(self, deployment_name, namespace):
-        """Delete a UAI deployment
+    def delete_job(self, job_name, namespace):
+        """Delete a UAI job
 
         """
         resp = None
         try:
             logger.info(
-                "delete deployment %s in namespace %s",
-                deployment_name,
+                "delete job %s in namespace %s",
+                job_name,
                 namespace
             )
-            resp = self.apps_v1.delete_namespaced_deployment(
-                name=deployment_name,
+            resp = self.batch_v1.delete_namespaced_job(
+                name=job_name,
                 namespace=namespace,
                 body=client.V1DeleteOptions(
                     propagation_policy='Background',
@@ -250,14 +237,14 @@ class UasBase:
         except ApiException as err:
             if err.status != 404:
                 logger.error(
-                    "Failed to delete deployment %s: %s",
-                    deployment_name,
+                    "Failed to delete job %s: %s",
+                    job_name,
                     err.reason
                 )
                 abort(
                     err.status,
-                    "Failed to delete deployment %s: %s" % (
-                        deployment_name,
+                    "Failed to delete job %s: %s" % (
+                        job_name,
                         err.reason
                     )
                 )
@@ -291,7 +278,7 @@ class UasBase:
         username = pod.metadata.labels.get("user", None)
         uai_name = pod.metadata.labels.get(
             "app",
-            "<internal error getting deployment name>"
+            "<internal error getting UAI name>"
         )
         opt_ports = pod.metadata.labels.get(
             "uas-uai-opt-ports",
@@ -351,7 +338,7 @@ class UasBase:
             uai_msg=uai_msg
         )
 
-    def get_pod_info(self, deployment_name):
+    def get_pod_info(self, job_name):
         """Retrieve pod information for a UAI pod from configuration.
 
         """
@@ -359,21 +346,21 @@ class UasBase:
         try:
             logger.info(
                 "getting pod info %s",
-                deployment_name
+                job_name
             )
             pod_resp = self.api.list_pod_for_all_namespaces(
-                label_selector="app=%s" % deployment_name,
+                label_selector="app=%s" % job_name,
             )
         except ApiException as err:
             logger.error(
                 "Failed to get pod info %s: %s",
-                deployment_name,
+                job_name,
                 err.reason
             )
             abort(
                 err.status,
                 "Failed to get pod info %s: %s" % (
-                    deployment_name,
+                    job_name,
                     err.reason
                 )
             )
@@ -384,8 +371,8 @@ class UasBase:
         if len(pod_resp.items) > 1:
             logger.warning(
                 "Oddly found more than one pod in "
-                "deployment %s",
-                deployment_name
+                "job %s",
+                job_name
             )
         # Only take the first one (there should only ever be one)
         pod = pod_resp.items[0]
@@ -395,11 +382,11 @@ class UasBase:
             logger.info(
                 "getting service info for %s-ssh in "
                 "namespace %s",
-                deployment_name,
+                job_name,
                 pod.metadata.namespace
             )
             srv_resp = self.api.read_namespaced_service(
-                name=deployment_name + "-ssh",
+                name=job_name + "-ssh",
                 namespace=pod.metadata.namespace
             )
         except ApiException as err:
@@ -407,13 +394,13 @@ class UasBase:
                 logger.error(
                     "Failed to get service info for "
                     "%s-ssh: %s",
-                    deployment_name,
+                    job_name,
                     err.reason
                 )
                 abort(
                     err.status,
                     "Failed to get service info for %s-ssh: %s" % (
-                        deployment_name,
+                        job_name,
                         err.reason
                     )
                 )
@@ -470,7 +457,7 @@ class UasBase:
 
         """
         service_name = uai_instance.get_service_name()
-        deployment = uai_instance.create_deployment_object(
+        job = uai_instance.create_job_object(
             uai_class=uai_class,
             uas_cfg=uas_cfg
         )
@@ -479,37 +466,34 @@ class UasBase:
             uai_class,
             uas_cfg
         )
-        # Make sure the UAI deployment is created
-        deploy_resp = None
+        # Make sure the UAI job is created
+        job_resp = None
         try:
             logger.info(
-                "getting deployment %s in namespace %s",
-                uai_instance.deployment_name,
+                "getting job %s in namespace %s",
+                uai_instance.job_name,
                 uai_class.namespace
             )
-            deploy_resp = self.apps_v1.read_namespaced_deployment(
-                uai_instance.deployment_name,
+            job_resp = self.batch_v1.read_namespaced_job(
+                uai_instance.job_name,
                 uai_class.namespace
             )
         except ApiException as err:
             if err.status != 404:
                 logger.error(
-                    "Failed to read deployment %s: %s",
-                    uai_instance.deployment_name,
+                    "Failed to read job %s: %s",
+                    uai_instance.job_name,
                     err.reason
                 )
                 abort(
                     err.status,
-                    "Failed to read deployment %s: %s" % (
-                        uai_instance.deployment_name,
+                    "Failed to read job %s: %s" % (
+                        uai_instance.job_name,
                         err.reason
                     )
                 )
-        if not deploy_resp:
-            deploy_resp = self.create_deployment(
-                deployment,
-                uai_class.namespace
-            )
+        if not job_resp:
+            job_resp = self.create_job(job, uai_class.namespace)
 
         # Start the UAI services
         logger.info("creating the UAI service %s", service_name)
@@ -519,12 +503,12 @@ class UasBase:
             uai_class.namespace
         )
         if not svc_resp:
-            # Clean up the deployment
+            # Clean up the UAI
             logger.error(
                 "failed to create service, deleting UAI %s",
-                uai_instance.deployment_name
+                uai_instance.job_name
             )
-            self.remove_uais([uai_instance.deployment_name])
+            self.remove_uais([uai_instance.job_name])
             abort(
                 404,
                 "Failed to create service: %s" % service_name
@@ -535,7 +519,7 @@ class UasBase:
         delay = 0.5
         while True:
             uai_info = self.get_pod_info(
-                deploy_resp.metadata.name
+                job_resp.metadata.name
             )
             if uai_info and uai_info.uai_ip:
                 break
@@ -552,7 +536,7 @@ class UasBase:
             )
         return uai_info
 
-    def select_deployments(self, labels=None, host=None):
+    def select_jobs(self, labels=None, host=None):
         """Get a list of UAI names from the specified host (if any) that meet
         the criteria in the specified labels (if any).
 
@@ -565,385 +549,76 @@ class UasBase:
             label_selector = "%s,%s" % (label_selector, ','.join(labels))
         try:
             logger.info(
-                "listing deployments matching: host %s,"
+                "listing jobs matching: host %s,"
                 " labels %s",
                 host,
                 label_selector
             )
-            resp = self.apps_v1.list_deployment_for_all_namespaces(
+            resp = self.batch_v1.list_job_for_all_namespaces(
                 label_selector=label_selector
             )
         except ApiException as err:
             if err.status != 404:
                 logger.error(
-                    "Failed to get deployment list: %s",
+                    "Failed to get job list: %s",
                     err.reason
                 )
-                abort(err.status, "Failed to get deployment list")
-        return [deployment.metadata.name for deployment in resp.items]
+                abort(err.status, "Failed to get job list")
+        return [job.metadata.name for job in resp.items]
 
-    def get_uai_namespace(self, deployment_name):
-        """Determine the namespace a named UAI deployment is deployed in.
+    def get_uai_namespace(self, job_name):
+        """Determine the namespace a named UAI is deployed in.
 
         """
-        resp = self.apps_v1.list_deployment_for_all_namespaces(
-            label_selector="app=%s" % deployment_name
+        resp = self.batch_v1.list_job_for_all_namespaces(
+            label_selector="app=%s" % job_name
         )
         if resp is None or not resp.items:
             return None
         if len(resp.items) > 1:
             logger.warning(
-                "Oddly found more than one deployment named %s",
-                deployment_name
+                "Oddly found more than one job named %s",
+                job_name
             )
         return resp.items[0].metadata.namespace
 
-    def get_uai_list(self, deploy_names):
+    def get_uai_list(self, job_names):
         """Get a list of UAIs from the specified host (if any)
         that meet the criteria in the specified label (if any).
 
         """
         uai_list = []
-        for deployment_name in deploy_names:
-            uai = self.get_pod_info(deployment_name)
+        for job_name in job_names:
+            uai = self.get_pod_info(job_name)
             if uai is not None:
                 uai_list.append(uai)
         return uai_list
 
-    def remove_uais(self, deploy_names):
-        """Remove a list of UAIs by their deployment names from the specified
+    def remove_uais(self, job_names):
+        """Remove a list of UAIs by their names from the specified
         namespace.
 
         """
         resp_list = []
-        for deployment_name in deploy_names:
-            namespace = self.get_uai_namespace(deployment_name)
+        for job_name in job_names:
+            namespace = self.get_uai_namespace(job_name)
             if namespace is None:
-                # This deployment doesn't exist or doesn't have a
-                # namespace (I dont think the latter is possible).
-                # Skip it.
+                # This job doesn't exist or doesn't have a namespace
+                # (I dont think the latter is possible).  Skip it.
                 continue
 
             # Do services first so that we don't orphan one if they abort
             service_resp = self.delete_service(
-                deployment_name + "-ssh",
+                job_name + "-ssh",
                 namespace
             )
-            deploy_resp = self.delete_deployment(
-                deployment_name,
+            job_resp = self.delete_job(
+                job_name,
                 namespace
             )
-            if deploy_resp is None and service_resp is None:
-                message = "Failed to delete %s - Not found" % deployment_name
+            if job_resp is None and service_resp is None:
+                message = "Failed to delete %s - Not found" % job_name
             else:
-                message = "Successfully deleted %s" % deployment_name
+                message = "Successfully deleted %s" % job_name
             resp_list.append(message)
         return resp_list
-
-
-class UAIInstance:
-    """This class carries information about individual UAI instances used
-    in creating UAIs that does not belong in a UAIClass.  It provides
-    a convenient container for information that is only known at UAI
-    creation time.
-
-    """
-    @staticmethod
-    def validate_ssh_key(ssh_key):
-        """
-        checks whether the ssh_key is a valid public key
-        ssh_key input is expected to be a string
-        :return: returns True if valid public key
-        :rtype bool
-        """
-        try:
-            ssh = sshpubkeys.SSHKey(ssh_key, strict=False,
-                                    skip_option_parsing=True)
-            ssh.parse()
-            return True
-        except (NotImplementedError, sshExceptions.MalformedDataError) as err:
-            UAS_AUTH_LOGGER.error("Invalid key: %s", err)
-        except sshExceptions.InvalidKeyError as err:
-            UAS_AUTH_LOGGER.error("Unknown key type: %s", err)
-        except Exception as err:  # pylint: disable=broad-except
-            UAS_AUTH_LOGGER.error("Invalid non-key input: %s", err)
-        return False
-
-    def get_public_key_str(self, public_key):
-        """Extract a public SSH key from its packing and verify that it looks
-        like a public SSH key (and not a private one in particular).
-
-        """
-        public_key_str = None
-        if public_key:
-            try:
-                # Depending on the API call, the public key may come
-                # in as an 'io.Bytes' object or as a string.  If it is
-                # an 'io.Bytes' object it needs to be turned into a
-                # string.  Otherwise, it can be used as is.
-                if isinstance(public_key, str):
-                    public_key_str = public_key
-                else:
-                    public_key_str = public_key.read().decode()
-                if not self.validate_ssh_key(public_key_str):
-                    # do not log the key here even if it's invalid, it
-                    # could be a private key accidentally passed in
-                    logger.info("create_uai - invalid ssh public key")
-                    abort(400, "Invalid ssh public key.")
-            except Exception:  # pylint: disable=broad-except
-                logger.info("create_uai - invalid ssh public key")
-                abort(400, "Invalid ssh public key.")
-        return public_key_str
-
-    def __init__(self, owner=None, public_key=None, passwd_str=None):
-        """Constructor
-
-        """
-        self.owner = owner
-        if isinstance(public_key, str):
-            self.public_key_str = public_key
-        else:
-            self.public_key_str = self.get_public_key_str(public_key)
-        self.passwd_str = passwd_str
-        dep_id = str(uuid.uuid4().hex[:8])
-        dep_owner = "no-owner" if owner is None else self.owner
-        self.deployment_name = 'uai-' + dep_owner + '-' + dep_id
-
-    def get_service_name(self):
-        """ Compute the service name of a UAI based on UAI parameters.
-
-        """
-        return self.deployment_name + "-ssh"
-
-    def get_env(self, uai_class=None):
-        """ Compute a K8s environment block for use in the UAI deployment
-
-        """
-        env = [
-            client.V1EnvVar(
-                name='UAS_NAME',
-                value=self.get_service_name()
-            ),
-            client.V1EnvVar(
-                name='UAS_PASSWD',
-                value=self.passwd_str
-            ),
-            client.V1EnvVar(
-                name='UAS_PUBKEY',
-                value=self.public_key_str
-            )
-        ]
-        if uai_class.uai_creation_class is not None:
-            env.append(
-                client.V1EnvVar(
-                    name='UAI_CREATION_CLASS',
-                    value=uai_class.uai_creation_class
-                )
-            )
-        return env
-
-    def gen_labels(self, uai_class=None):
-        """Generate labels for a UAI Deployment
-
-        """
-        ret = {
-            "app": self.deployment_name,
-            "uas": "managed"
-        }
-        if self.owner is not None:
-            ret['user'] = self.owner
-        if uai_class is not None:
-            if uai_class.uai_creation_class is not None:
-                ret['uas-uai-creation-class'] = uai_class.uai_creation_class
-            if uai_class.opt_ports is not None:
-                ret['uas-uai-opt-ports'] = "-".join(uai_class.opt_ports)
-            ret['uas-public-ip'] = str(uai_class.public_ip)
-            ret['uas-class-id'] = uai_class.class_id
-        return ret
-
-    # pylint: disable=too-many-locals
-    def create_deployment_object(self, uai_class, uas_cfg):
-        """Construct a deployment for a UAI or Broker
-
-        """
-        # If we are using macvlan then we will set that up in an
-        # annotation in the metadata of the deployment, otherwise, the
-        # annotations will be None. USE_MACVLAN is based on
-        # configuration from the Helm chart that can be set at service
-        # deployment time.  We only set this up in UAIs made from a
-        # class that also has the 'uai_compute_network' flag turned
-        # on since some UAIs don't want compute network connectivity.
-        meta_annotations = None
-        if os.environ.get('USE_MACVLAN', 'true').lower() == 'true' and \
-           uai_class.uai_compute_network:
-            meta_annotations = {
-                'k8s.v1.cni.cncf.io/networks': 'macvlan-uas-nmn-conf@nmn1'
-            }
-
-        pod_metadata = client.V1ObjectMeta(
-            labels=self.gen_labels(uai_class),
-            annotations=meta_annotations
-        )
-        deploy_metadata = client.V1ObjectMeta(
-            name=self.deployment_name,
-            labels=self.gen_labels(uai_class)
-        )
-        volume_list = uai_class.volume_list
-        resources = None
-        if uai_class.resource_id is not None:
-            resources = {}
-            limit_json = UAIResource.get(uai_class.resource_id).limit
-            request_json = UAIResource.get(uai_class.resource_id).request
-            if limit_json:
-                resources['limits'] = json.loads(limit_json)
-            if request_json:
-                resources['requests'] = json.loads(request_json)
-        if not resources:
-            resources = None
-        container_ports = uas_cfg.gen_port_list(
-            service=False,
-            opt_ports=[
-                int(port) for port in uai_class.opt_ports
-            ] if uai_class.opt_ports is not None else None
-        )
-        logger.info(
-            "UAI Name: %s; Container ports: %s; Optional ports: %s",
-            self.deployment_name,
-            container_ports,
-            uai_class.opt_ports
-        )
-
-        # Configure Pod template container
-        container = client.V1Container(
-            name=self.deployment_name,
-            image=UAIImage.get(uai_class.image_id).imagename,
-            resources=resources,
-            env=self.get_env(uai_class),
-            ports=container_ports,
-            volume_mounts=uas_cfg.gen_volume_mounts(volume_list),
-            readiness_probe=uas_cfg.create_readiness_probe()
-        )
-
-        # Create and configure affinity
-        node_selector_terms = [
-            client.V1NodeSelectorTerm(
-                match_expressions=[
-                    client.V1NodeSelectorRequirement(
-                        key='node-role.kubernetes.io/master',
-                        operator='DoesNotExist'
-                    ),
-                    client.V1NodeSelectorRequirement(
-                        key='uas',
-                        operator='NotIn',
-                        values=['False', 'false', 'FALSE']
-                    )
-                ]
-            )
-        ]
-        node_selector = client.V1NodeSelector(node_selector_terms)
-        node_affinity = client.V1NodeAffinity(
-            required_during_scheduling_ignored_during_execution=node_selector
-        )
-        affinity = client.V1Affinity(node_affinity=node_affinity)
-        # pylint: disable=unnecessary-comprehension
-        tolerations = [toleration for toleration in BASE_UAI_TOLERATIONS]
-        if uai_class.tolerations is not None:
-            toleration_list = json.loads(uai_class.tolerations)
-            for toleration in toleration_list:
-                tolerations.append(client.V1Toleration(**toleration))
-
-        priority_class_name = 'uai-priority'
-        if uai_class.priority_class_name is not None:
-            priority_class_name = uai_class.priority_class_name
-
-        template = client.V1PodTemplateSpec(
-            metadata=pod_metadata,
-            spec=client.V1PodSpec(
-                priority_class_name=priority_class_name,
-                containers=[container],
-                affinity=affinity,
-                volumes=uas_cfg.gen_volumes(volume_list),
-                tolerations=tolerations
-            )
-        )
-
-        # Create the specification of deployment
-        spec = client.V1DeploymentSpec(
-            replicas=1,
-            selector={
-                'matchLabels': {
-                    'app': self.deployment_name
-                }
-            },
-            template=template
-        )
-        # Instantiate the deployment object
-        deployment = client.V1Deployment(
-            api_version="apps/v1",
-            kind="Deployment",
-            metadata=deploy_metadata,
-            spec=spec
-        )
-        return deployment
-
-    def create_service_object(self, uai_class, uas_cfg):
-        """
-        Create a service object for the deployment of the UAI.
-
-        """
-        # Pick the service type based on the value of 'public_ip' in
-        # the UAI Class.  This is a lot simpler than it looks if you
-        # delve into it, but I am using the code that was here to do
-        # this. That code bases the service class (SSH point of
-        # access) on two strings: "service" (which basically means an
-        # internal ClusterIP) and "ssh" (which basically means a
-        # LoadBalncer IP or a NodePort).  Instead of reworking all
-        # that logic, I am picking one or the other here based on
-        # whether 'public_ip' is true or false.
-        service_type = "ssh" if uai_class.public_ip else "service"
-        metadata = client.V1ObjectMeta(
-            name=self.get_service_name(),
-            labels=self.gen_labels(uai_class),
-        )
-        ports = uas_cfg.gen_port_list(
-            service_type,
-            service=True,
-            opt_ports=[
-                int(port) for port in uai_class.opt_ports
-            ] if uai_class.opt_ports is not None else None
-        )
-
-        # svc_type is a dict with the following fields:
-        #   'svc_type': (NodePort, ClusterIP, or LoadBalancer)
-        #   'ip_pool': (None, or a specific pool)  Valid only for LoadBalancer.
-        #   'valid': (True or False) is svc_type is valid or not
-        svc_type = uas_cfg.get_svc_type(service_type)
-        if not svc_type['valid']:
-            # Invalid svc_type given.
-            msg = (
-                "Unsupported service type '{}' configured, "
-                "contact sysadmin. Valid service types are "
-                "NodePort, ClusterIP, and LoadBalancer.".format(
-                    svc_type['svc_type']
-                )
-            )
-            abort(400, msg)
-        # Check if LoadBalancer and whether an IP pool is set
-        if svc_type['svc_type'] == "LoadBalancer" and svc_type['ip_pool']:
-            # A specific IP pool is given, update the metadata with
-            # annotations
-            metadata.annotations = {
-                "metallb.universe.tf/address-pool": svc_type['ip_pool']
-            }
-        spec = client.V1ServiceSpec(
-            selector={'app': self.deployment_name},
-            type=svc_type['svc_type'],
-            ports=ports
-        )
-        service = client.V1Service(
-            api_version="v1",
-            kind="Service",
-            metadata=metadata,
-            spec=spec
-        )
-        return service
